+++ scripts/run_examples.sh
@@ -16,6 +16,7 @@ basics=(
   jacobian_hessian
   lie_groups
   linear_algebra
+  model_ingestion
   ode
   optimization_solvers
   root_finding